feat: content -> block 네이밍 변경
