+++ backend/src/main/java/org/donggle/backend/application/repository/BlockRepository.java
@@ -1,6 +1,6 @@
 package org.donggle.backend.application.repository;
 
-import org.donggle.backend.domain.writing.content.Block;
+import org.donggle.backend.domain.writing.block.Block;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
