+++ backend/src/main/java/org/donggle/backend/domain/parser/markdown/MarkDownParser.java
@@ -4,16 +4,16 @@ import lombok.RequiredArgsConstructor;
 import org.donggle.backend.domain.writing.BlockType;
 import org.donggle.backend.domain.writing.Style;
 import org.donggle.backend.domain.writing.Writing;
-import org.donggle.backend.domain.writing.content.Block;
-import org.donggle.backend.domain.writing.content.CodeBlock;
-import org.donggle.backend.domain.writing.content.Depth;
-import org.donggle.backend.domain.writing.content.HorizontalRulesBlock;
-import org.donggle.backend.domain.writing.content.ImageBlock;
-import org.donggle.backend.domain.writing.content.ImageCaption;
-import org.donggle.backend.domain.writing.content.ImageUrl;
-import org.donggle.backend.domain.writing.content.Language;
-import org.donggle.backend.domain.writing.content.NormalBlock;
-import org.donggle.backend.domain.writing.content.RawText;
+import org.donggle.backend.domain.writing.block.Block;
+import org.donggle.backend.domain.writing.block.CodeBlock;
+import org.donggle.backend.domain.writing.block.Depth;
+import org.donggle.backend.domain.writing.block.HorizontalRulesBlock;
+import org.donggle.backend.domain.writing.block.ImageBlock;
+import org.donggle.backend.domain.writing.block.ImageCaption;
+import org.donggle.backend.domain.writing.block.ImageUrl;
+import org.donggle.backend.domain.writing.block.Language;
+import org.donggle.backend.domain.writing.block.NormalBlock;
+import org.donggle.backend.domain.writing.block.RawText;
 
 import java.util.ArrayList;
 import java.util.Arrays;
