+++ backend/src/main/java/org/donggle/backend/domain/renderer/html/HtmlRenderer.java
@@ -1,11 +1,11 @@
 package org.donggle.backend.domain.renderer.html;
 
 import org.donggle.backend.domain.writing.BlockType;
-import org.donggle.backend.domain.writing.content.Block;
-import org.donggle.backend.domain.writing.content.CodeBlock;
-import org.donggle.backend.domain.writing.content.HorizontalRulesBlock;
-import org.donggle.backend.domain.writing.content.ImageBlock;
-import org.donggle.backend.domain.writing.content.NormalBlock;
+import org.donggle.backend.domain.writing.block.Block;
+import org.donggle.backend.domain.writing.block.CodeBlock;
+import org.donggle.backend.domain.writing.block.HorizontalRulesBlock;
+import org.donggle.backend.domain.writing.block.ImageBlock;
+import org.donggle.backend.domain.writing.block.NormalBlock;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -21,34 +21,33 @@ public class HtmlRenderer {
 
     public String render(final List<Block> blocks) {
         final StringBuilder result = new StringBuilder();
-        final List<NormalBlock> subContent = new ArrayList<>();
+        final List<NormalBlock> subBlock = new ArrayList<>();
         String htmlText;
 
         for (final Block block : blocks) {
-            htmlText = createHtmlText(subContent, block);
-            if (!htmlText.isBlank() && !subContent.isEmpty()) {
-                result.append(renderList(subContent));
-                subContent.clear();
+            htmlText = createHtmlText(subBlock, block);
+            if (!htmlText.isBlank() && !subBlock.isEmpty()) {
+                result.append(renderList(subBlock));
+                subBlock.clear();
             }
-
             result.append(htmlText);
         }
 
-        if (!subContent.isEmpty()) {
-            result.append(renderList(subContent));
+        if (!subBlock.isEmpty()) {
+            result.append(renderList(subBlock));
         }
 
         return result.toString();
     }
 
-    private String createHtmlText(final List<NormalBlock> subContent, final Block block) {
+    private String createHtmlText(final List<NormalBlock> subBlock, final Block block) {
         final BlockType blockType = block.getBlockType();
         String htmlText = "";
 
         switch (blockType) {
             case HEADING1, HEADING2, HEADING3, HEADING4, HEADING5, HEADING6, PARAGRAPH, BLOCKQUOTE ->
-                    htmlText = renderNormalContent((NormalBlock) block);
-            case ORDERED_LIST, UNORDERED_LIST -> subContent.add((NormalBlock) block);
+                    htmlText = renderNormalBlock((NormalBlock) block);
+            case ORDERED_LIST, UNORDERED_LIST -> subBlock.add((NormalBlock) block);
             case CODE_BLOCK -> htmlText = renderCodeBlock((CodeBlock) block);
             case IMAGE -> htmlText = renderImage((ImageBlock) block);
             case HORIZONTAL_RULES -> htmlText = renderHorizontalRules((HorizontalRulesBlock) block);
@@ -56,10 +55,10 @@ public class HtmlRenderer {
         return htmlText;
     }
 
-    private String renderNormalContent(final NormalBlock content) {
-        final HtmlType htmlType = HtmlType.findByBlockType(content.getBlockType());
-        final String depth = renderDepth(content.getDepthValue());
-        final String rawText = htmlStyleRenderer.render(content.getRawTextValue(), content.getStyles());
+    private String renderNormalBlock(final NormalBlock block) {
+        final HtmlType htmlType = HtmlType.findByBlockType(block.getBlockType());
+        final String depth = renderDepth(block.getDepthValue());
+        final String rawText = htmlStyleRenderer.render(block.getRawTextValue(), block.getStyles());
 
         return htmlType.getStartTag() + depth + rawText + htmlType.getEndTag();
     }
@@ -68,10 +67,10 @@ public class HtmlRenderer {
         return HTML_TAB.repeat(Math.max(0, depth));
     }
 
-    private String renderCodeBlock(final CodeBlock content) {
-        final HtmlType htmlType = HtmlType.findByBlockType(content.getBlockType());
-        final String language = content.getLanguageValue();
-        final String rawText = convertToEscape(content.getRawTextValue());
+    private String renderCodeBlock(final CodeBlock block) {
+        final HtmlType htmlType = HtmlType.findByBlockType(block.getBlockType());
+        final String language = block.getLanguageValue();
+        final String rawText = convertToEscape(block.getRawTextValue());
 
         final String startTag = htmlType.getStartTag()
                 .replace("${language}", language);
@@ -110,67 +109,67 @@ public class HtmlRenderer {
         }
     }
 
-    private String renderList(final List<NormalBlock> contents) {
+    private String renderList(final List<NormalBlock> blocks) {
         final StringBuilder result = new StringBuilder();
-        final NormalBlock firstContent = contents.get(0);
-        final NormalBlock endContent = contents.get(contents.size() - 1);
+        final NormalBlock firstBlock = blocks.get(0);
+        final NormalBlock endBlock = blocks.get(blocks.size() - 1);
 
-        addFirstHtmlType(firstContent, result);
-        addInnerHtmlType(contents, result);
-        addEndHtmlType(endContent, result);
+        addFirstHtmlType(firstBlock, result);
+        addInnerHtmlType(blocks, result);
+        addEndHtmlType(endBlock, result);
 
         return result.toString();
     }
 
-    private void addFirstHtmlType(final NormalBlock content, final StringBuilder result) {
-        final HtmlType htmlType = HtmlType.findByBlockType(content.getBlockType());
+    private void addFirstHtmlType(final NormalBlock block, final StringBuilder result) {
+        final HtmlType htmlType = HtmlType.findByBlockType(block.getBlockType());
         result.append(htmlType.getStartTag());
     }
 
-    private void addInnerHtmlType(final List<NormalBlock> contents, final StringBuilder result) {
-        final int contentSize = contents.size();
+    private void addInnerHtmlType(final List<NormalBlock> blocks, final StringBuilder result) {
+        final int blockSize = blocks.size();
 
-        for (int i = 0; i < contentSize - 1; i++) {
-            final NormalBlock currentContent = contents.get(i);
-            final NormalBlock nextContent = contents.get(i + 1);
+        for (int i = 0; i < blockSize - 1; i++) {
+            final NormalBlock currentBlock = blocks.get(i);
+            final NormalBlock nextBlock = blocks.get(i + 1);
 
-            result.append(renderLine(currentContent, nextContent));
+            result.append(renderLine(currentBlock, nextBlock));
         }
     }
 
-    private String renderLine(final NormalBlock content, final NormalBlock nextContent) {
-        final String rawText = htmlStyleRenderer.render(content.getRawTextValue(), content.getStyles());
+    private String renderLine(final NormalBlock block, final NormalBlock nextBlock) {
+        final String rawText = htmlStyleRenderer.render(block.getRawTextValue(), block.getStyles());
         final String line = HtmlType.LIST.getStartTag() + rawText + HtmlType.LIST.getEndTag();
 
-        if (content.getDepthValue() > nextContent.getDepthValue()) {
-            final HtmlType htmlType = HtmlType.findByBlockType(content.getBlockType());
+        if (block.getDepthValue() > nextBlock.getDepthValue()) {
+            final HtmlType htmlType = HtmlType.findByBlockType(block.getBlockType());
             return line + htmlType.getEndTag();
         }
 
-        if (content.getDepthValue() == nextContent.getDepthValue()) {
-            if (content.getBlockType().equals(nextContent.getBlockType())) {
+        if (block.getDepthValue() == nextBlock.getDepthValue()) {
+            if (block.getBlockType().equals(nextBlock.getBlockType())) {
                 return line;
             }
-            final HtmlType htmlType = HtmlType.findByBlockType(content.getBlockType());
-            final HtmlType nextHtmlType = HtmlType.findByBlockType(nextContent.getBlockType());
+            final HtmlType htmlType = HtmlType.findByBlockType(block.getBlockType());
+            final HtmlType nextHtmlType = HtmlType.findByBlockType(nextBlock.getBlockType());
             return line + htmlType.getEndTag() + nextHtmlType.getStartTag();
         }
 
-        final HtmlType nextHtmlType = HtmlType.findByBlockType(nextContent.getBlockType());
+        final HtmlType nextHtmlType = HtmlType.findByBlockType(nextBlock.getBlockType());
         return line + nextHtmlType.getStartTag();
     }
 
-    private void addEndHtmlType(final NormalBlock content, final StringBuilder result) {
-        final HtmlType htmlType = HtmlType.findByBlockType(content.getBlockType());
-        final String rawText = htmlStyleRenderer.render(content.getRawTextValue(), content.getStyles());
+    private void addEndHtmlType(final NormalBlock block, final StringBuilder result) {
+        final HtmlType htmlType = HtmlType.findByBlockType(block.getBlockType());
+        final String rawText = htmlStyleRenderer.render(block.getRawTextValue(), block.getStyles());
         final String line = HtmlType.LIST.getStartTag() + rawText + HtmlType.LIST.getEndTag() + htmlType.getEndTag();
         result.append(line);
     }
 
-    private String renderImage(final ImageBlock content) {
-        final HtmlType htmlType = HtmlType.findByBlockType(content.getBlockType());
-        final String caption = content.getImageCaptionValue();
-        final String url = content.getImageUrlValue();
+    private String renderImage(final ImageBlock block) {
+        final HtmlType htmlType = HtmlType.findByBlockType(block.getBlockType());
+        final String caption = block.getImageCaptionValue();
+        final String url = block.getImageUrlValue();
 
         final String startTag = htmlType.getStartTag()
                 .replace("caption", caption)
