+++ backend/src/main/java/org/donggle/backend/domain/writing/block/Block.java
@@ -1,4 +1,4 @@
-package org.donggle.backend.domain.writing.content;
+package org.donggle.backend.domain.writing.block;
 
 import jakarta.persistence.DiscriminatorColumn;
 import jakarta.persistence.Embedded;
