+++ backend/src/main/java/org/donggle/backend/domain/writing/block/NormalBlock.java
@@ -1,4 +1,4 @@
-package org.donggle.backend.domain.writing.content;
+package org.donggle.backend.domain.writing.block;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
