+++ backend/src/test/java/org/donggle/backend/domain/renderer/html/HtmlRendererTest.java
@@ -6,13 +6,13 @@ import org.donggle.backend.domain.member.MemberName;
 import org.donggle.backend.domain.writing.BlockType;
 import org.donggle.backend.domain.writing.Title;
 import org.donggle.backend.domain.writing.Writing;
-import org.donggle.backend.domain.writing.content.Block;
-import org.donggle.backend.domain.writing.content.CodeBlock;
-import org.donggle.backend.domain.writing.content.Depth;
-import org.donggle.backend.domain.writing.content.HorizontalRulesBlock;
-import org.donggle.backend.domain.writing.content.Language;
-import org.donggle.backend.domain.writing.content.NormalBlock;
-import org.donggle.backend.domain.writing.content.RawText;
+import org.donggle.backend.domain.writing.block.Block;
+import org.donggle.backend.domain.writing.block.CodeBlock;
+import org.donggle.backend.domain.writing.block.Depth;
+import org.donggle.backend.domain.writing.block.HorizontalRulesBlock;
+import org.donggle.backend.domain.writing.block.Language;
+import org.donggle.backend.domain.writing.block.NormalBlock;
+import org.donggle.backend.domain.writing.block.RawText;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
