+++ backend/src/test/java/org/donggle/backend/domain/writing/BlockTest.java
@@ -7,10 +7,10 @@ import org.donggle.backend.application.repository.WritingRepository;
 import org.donggle.backend.domain.category.Category;
 import org.donggle.backend.domain.member.Member;
 import org.donggle.backend.domain.member.MemberName;
-import org.donggle.backend.domain.writing.content.Block;
-import org.donggle.backend.domain.writing.content.CodeBlock;
-import org.donggle.backend.domain.writing.content.Language;
-import org.donggle.backend.domain.writing.content.RawText;
+import org.donggle.backend.domain.writing.block.Block;
+import org.donggle.backend.domain.writing.block.CodeBlock;
+import org.donggle.backend.domain.writing.block.Language;
+import org.donggle.backend.domain.writing.block.RawText;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
